feat: RefreshButton 버튼 분리 및 위치 이동
