+++ packages/view/src/App.scss
@@ -9,10 +9,22 @@ body {
 }
 
 .header-container {
+  position: relative;
   display: flex;
   height: 60px;
   justify-content: center;
   align-items: center;
+}
+
+.header-buttons {
+  position: absolute;
+  right: 0;
+  display: flex;
+  align-items: center;
+
+  > *:not(:last-child) {
+    margin-right: 4px;
+  }
 }
 
 .top-container {
