+++ packages/view/src/components/RefreshButton/index.ts
@@ -1,0 +1,1 @@
+export { default as RefreshButton } from "./RefreshButton";
