+++ packages/view/src/components/TemporalFilter/TemporalFilter.scss
@@ -2,30 +2,6 @@
   display: flex;
   flex-direction: column;
 
-  .data-control-container {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    position: fixed;
-    width: 100%;
-    padding: 1px;
-    // background: #212121;
-    color: white;
-    height: 30px;
-
-    .filter {
-      display: flex;
-      flex: 1;
-      justify-content: center;
-      align-items: center;
-    }
-
-    .theme-selector {
-      margin-right: 3px;
-    }
-  }
-
   .line-charts {
     display: flex;
     flex-direction: column;
@@ -77,28 +53,4 @@
   font-size: 10px;
   font-weight: 500;
   fill: white;
-}
-
-.refresh-button {
-  width: 2.2rem;
-  height: 1.2rem;
-  position: absolute;
-  right: 0;
-  border: 0px solid white;
-  border-radius: 5px;
-  color: white;
-  background-color: transparent;
-  cursor: pointer;
-  padding-right: 2.6rem;
-  z-index: 10;
-
-  &:active {
-    background-color: black;
-    color: white;
-  }
-
-  &:hover {
-    background-color: white;
-    color: black;
-  }
 }
