+++ packages/view/src/components/ThemeSelector/ThemeSelector.scss
@@ -1,16 +1,13 @@
 @import "styles/_pallete";
 
 .theme-selector {
-    display: flex;
-    position: absolute;
-    right: 3rem;
-    input[type=color] {
-        outline: none;
-        -webkit-appearance: none;
-        padding: 0;
-        border: none;
-        background-color: transparent;
-        width: 54px;
-        height: 30px;
-    }
+  input[type="color"] {
+    outline: none;
+    -webkit-appearance: none;
+    padding: 0;
+    border: none;
+    background-color: transparent;
+    width: 54px;
+    height: 30px;
+  }
 }
